requirements.txt
Flask>=2.0.0,<3.0.0
Flask-CORS>=3.0.0,<4.0.0
python-dotenv>=0.19.0,<1.0.0
requests>=2.25.0,<3.0.0
openai>=1.0.0,<2.0.0
gunicorn>=20.1.0,<21.0.0 